--- alert_text.py
@@ -1,27 +1,0 @@
-def get_participant_notification_text(last_name, first_name, login, password):
-    mail_text2 = f"""Здравствуйте, {last_name.capitalize()} {first_name.capitalize()}!  
-
-Поздравляем, Вы оплатили абонемент на месяц совместных занятий в онлайн-формате "Друзья Школы Гивина". 
-
-Ваш zoom-аккаунт:
-Логин: {login}
-Пароль: {password}
-
-Сохраните себе эти данные, чтобы не потерять их. 
-
-Эти данные вы можете использовать с настоящего момента:
-1) Скачайте приложение Zoom на компьютер, если ещё не cделали это ранее. 
-2) Установите приложение Zoom на ваш компьютер.
-3) Запустите эту программу.
-4) Нажмите кнопку Sign In ("Войти в..").
-5) Введите логин и пароль, предоставленные вам в этом письме .
-6) Поставьте птичку (галку) в поле Keep me logged in ("Не выходить из системы").
-7) Нажмите Sign In ("Войти"). 
-8) Далее из чата Объявлений в телеграмме найдет сообщение с ссылкой на занятия. 
-   Нажмите на неё. Она будет открываться в браузере, появится сверху сообщение с кнопкой, 
-   жмём на кнопку Open ZOOM Meetings (либо Открыть ZOOM)
-9) Появится окно для ввода пароля конференции. Здесь вводим три цифры 355. 
-
-С благодарностью и сердечным теплом,
-команда Школы Гивина."""
-    return mail_text2
